WIP: tighten trip editor spike validation cleanup (checkpoint)
