+++ package.json
@@ -8,7 +8,7 @@
   },
   "dependencies": {
     "@vitejs/plugin-vue": "6.0.3",
-    "vite": "7.3.0",
+    "vite": "7.3.3",
     "typescript": "5.9.3",
     "vue": "3.5.34",
     "leaflet": "1.9.4",
